remove Proxydead tests
